refactor: move related to webhook module enums to the module itself
