+++ packages/apps/reputation-oracle/server/src/modules/webhook/types.ts
@@ -1,0 +1,21 @@
+export enum IncomingWebhookEventType {
+  JOB_COMPLETED = 'job_completed',
+}
+
+export enum OutgoingWebhookEventType {
+  ESCROW_COMPLETED = 'escrow_completed',
+  ABUSE_DETECTED = 'abuse_detected',
+  ABUSE_DISMISSED = 'abuse_dismissed',
+}
+
+export enum IncomingWebhookStatus {
+  PENDING = 'pending',
+  COMPLETED = 'completed',
+  FAILED = 'failed',
+}
+
+export enum OutgoingWebhookStatus {
+  PENDING = 'pending',
+  SENT = 'sent',
+  FAILED = 'failed',
+}
