+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-incoming.entity.ts
@@ -1,8 +1,8 @@
 import { Column, Entity, Index } from 'typeorm';
 
 import { DATABASE_SCHEMA_NAME } from '../../common/constants';
-import { IncomingWebhookStatus } from '../../common/enums';
 import { BaseEntity } from '../../database/base.entity';
+import { IncomingWebhookStatus } from './types';
 
 @Entity({ schema: DATABASE_SCHEMA_NAME, name: 'webhook_incoming' })
 @Index(['chainId', 'escrowAddress'], { unique: true })
