+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-outgoing.repository.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { DataSource, LessThanOrEqual } from 'typeorm';
 
 import { ServerConfigService } from '../../config/server-config.service';
-import { OutgoingWebhookStatus } from '../../common/enums/webhook';
 import { BaseRepository } from '../../database/base.repository';
+import { OutgoingWebhookStatus } from './types';
 import { OutgoingWebhookEntity } from './webhook-outgoing.entity';
 
 @Injectable()
