+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook.dto.ts
@@ -2,8 +2,8 @@ import { ChainId } from '@human-protocol/sdk';
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { IsEthereumAddress, IsObject, IsOptional } from 'class-validator';
 
-import { IncomingWebhookEventType } from '../../common/enums';
 import { IsChainId, IsLowercasedEnum } from '../../common/validators';
+import { IncomingWebhookEventType } from './types';
 
 export class IncomingWebhookDto {
   @ApiProperty({ name: 'chain_id' })
